\name{TERGMundir-package}
\alias{TERGMundir-package}
\alias{TERGMundir}
\docType{package}
\title{
\packageTitle{TERGMundir}
}
\description{
\packageDescription{TERGMundir}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{TERGMundir}
\packageIndices{TERGMundir}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{TERGMundir}

Maintainer: \packageMaintainer{TERGMundir}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
